+++ src/control_plane/workload/mod.rs
@@ -9,6 +9,9 @@ mod project_process_plan;
 mod project_process_plan_options;
 mod project_process_request;
 mod project_process_request_options;
+mod runtime_environment;
+mod runtime_environment_error;
+mod runtime_environment_options;
 mod validate_image_digest;
 mod workload_plan_error;
 
@@ -20,4 +23,7 @@ pub(crate) use project_process_plan::ProjectProcessPlan;
 pub(crate) use project_process_plan_options::ProjectProcessPlanOptions;
 pub(crate) use project_process_request::project_process_request;
 pub(crate) use project_process_request_options::ProjectProcessRequestOptions;
+pub(crate) use runtime_environment::RuntimeEnvironment;
+pub(crate) use runtime_environment_error::RuntimeEnvironmentError;
+pub(crate) use runtime_environment_options::RuntimeEnvironmentOptions;
 pub(crate) use workload_plan_error::WorkloadPlanError;